requirements.txt
openai-whisper
pydub
numpy>=1.24,<2.0
transformers<=4.44.2
torch>=2.1,<2.9
torchaudio>=2.1,<2.9
gradio
groq
python-dotenv
TTS==0.22.0
